backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
numpy==1.24.3
SQLAlchemy==2.0.15
Werkzeug==2.3.6
redis==4.5.4
celery==5.3.0
